+++ sakura_core/charset/ShiftJis.cpp
@@ -22,7 +22,7 @@
 	@note nIdxは予め文字の先頭位置とわかっていなければならない．
 	2バイト文字の2バイト目をnIdxに与えると正しい結果が得られない．
 */
-int ShiftJis::GetSizeOfChar(const char* pData, int nDataLen, int nIdx)
+size_t ShiftJis::GetSizeOfChar(const char* pData, size_t nDataLen, size_t nIdx)
 {
 	if (nIdx >= nDataLen) {
 		return 0;
@@ -114,7 +114,7 @@ CodeConvertResult ShiftJis::SJISToUnicode(
 	)
 {
 	// ソース取得
-	int nSrcLen;
+	size_t nSrcLen;
 	const char* pSrc = reinterpret_cast<const char*>(src.GetRawPtr(&nSrcLen));
 	if (nSrcLen == 0) {
 		pDstMem->_SetStringLength(0);
